+++ app/src/main/java/com/betterlife/antifragile/data/remote/MemberApiService.kt
@@ -38,16 +38,6 @@ interface MemberApiService {
     @GET("/members/remain-recommend-number")
     suspend fun getRemainRecommendNumber(): Response<BaseResponse<MemberRemainNumberResponse>>
 
-    @GET("/members/duplication-check")
-    suspend fun checkNicknameExistence(
-        @Query("nickname") nickname: String
-    ): Response<BaseResponse<NicknameDuplicateResponse>>
-
-    @GET("/members/duplication-check")
-    suspend fun checkMemberStatus(
-        @Query("nickname") nickname: String
-    ): Response<BaseResponse<NicknameDuplicateResponse>>
-
     @DELETE("/auth")
     suspend fun delete(): Response<BaseResponse<Any?>>
