+++ app/src/main/java/com/betterlife/antifragile/data/repository/AuthRepository.kt
@@ -9,6 +9,7 @@ import com.betterlife.antifragile.data.model.auth.response.AuthSignUpResponse
 import com.betterlife.antifragile.data.model.base.BaseResponse
 import com.betterlife.antifragile.data.model.enums.LoginType
 import com.betterlife.antifragile.data.model.member.response.MemberExistenceResponse
+import com.betterlife.antifragile.data.model.member.response.NicknameDuplicateResponse
 import com.betterlife.antifragile.data.remote.AuthApiService
 import okhttp3.MultipartBody
 
@@ -35,6 +36,14 @@ class AuthRepository(
     ): BaseResponse<MemberExistenceResponse> {
         return safeApiCall {
             authApiService.checkMemberExistence(email, loginType)
+        }
+    }
+
+    suspend fun checkNicknameExistence(
+        nickname: String
+    ): BaseResponse<NicknameDuplicateResponse> {
+        return safeApiCall {
+            authApiService.checkNicknameExistence(nickname)
         }
     }
